update train.py and inference.py
