+++ running.sh
@@ -1,7 +1,10 @@
+# comman example
 python train.py \
 --do_train \
 --fp16 \
---model_type rbert \
+--PLM uclanlp/plbart-base \
+--model_category plbart \
+--model_name BartEncoderConcatModel \
 --logging_strategy steps \
 --logging_steps 500 \
 --evaluation_strategy steps \
@@ -17,15 +20,18 @@ python train.py \
 --per_device_train_batch_size 16 \
 --per_device_eval_batch_size 32 \
 --gradient_accumulation_steps 2 \
+--eval_accumulation_steps 4 \
 --load_best_model_at_end \
 --metric_for_best_model accuracy \
 --warmup_ratio 0.05 \
 --weight_decay 1e-2
 
 # python inference.py \
-# --tokenizer microsoft/codebert-base \
+# --tokenizer uclanlp/plbart-base \
+# --model_category plbart \
+# --model_name BartEncoderConcatModel \
 # --output_dir results \
-# --file_name codebert_rbert_trainer_EP:2_BS:32_WR:0.05_WD:1e-2_LR:2e-5.csv \
+# --file_name plbart_base_encoderconcat_EP:2_BS:32_WR:0.05_WD:1e-2_LR:2e-5.csv \
 # --PLM ./checkpoints \
 # --fp16 \
 # --per_device_eval_batch_size 32
